Begin complete refactor of site, implement mobile first header
This commit begins a total refactor of all code on the site, beginning
with removal of superflous html and the commenting out of all code in
the style.css file.

Styles have also been reintroduced for the header feature. In doing so
they have been completely re-factored into a mobile-first layout.
Default styles are targeted at the mobile layout, while tablet and web
layouts are targeted in subsequent media queries.

An additional media query has been added to allow for a custom
breakpoint for the mobile navbar, which now cleanly breaks with the
tablet laout.