Add refactored card and responsive images
This commit adds one card per page back into the html files with a
standardised, less cluttered layout, built in a mobile first
methodology. Cards primarily use margins and padding to acheive layout
as opposed to heavily relying on grid.

The column edges are now more defined due to a choice to incorporate a
justfified text layout, as well as the centering of the read more
button. Additionally the card headings now have margins which align with
these edges, which is most notable on the history layout.

Images now use the <picture> tag to achieve responsive serving of images
appropriate to the device they are being viewed on, in line with the
media query breakpoints. A number of images have been flagged as
requiring updates on the design side of things, and will be dealt with
in due course.