Update Assets In The img Directory
This commit updates all assets in the img folder, correcting both some
cropping faults as well as some colour casts that were absent. In
addition there is now one image for each screen size, simplifying the
process of delivering content to each device.

Visible images in the html files have been updated. However images that
are currently commented out with be dealt with when they are added
back into the flow of the document.