Add Full HTML Content Of Primary Pages
This commit reintroduces the remaining primary content on the main html
pages. This includes all remaining cards, images, and text.

Images have again been added using the picture tag and the srcset
attribute to allow the serving of images appropriate to the screen size.

Currently the content renders well in a mobile view, consistent with the
mobile first approach. This will be tailored to tablet and web devices
in a later commit.