Reintroduce Styles For Instruments & Dances Pages
This commit reintroduces styles on both the instruments.html and
dances.html pages which were broken in the previous commit.

In line with the mobile first appraoch, this commit produces a clean
layout across all break points, and at all screen widths in between.

Images are delivered via the picture element which in particular allows
the cards on these two pages to display perfectly without any warping of
the images, as in the pre-refactor layout.

Note - The history page layout is still broken by this commit and will
be addressed in the next commit.