Add Refactored Layout For History Page
This commit reintroduces styles for the history.html page in a mobile
first responsive layout. Secondary images are hidden on all except the
web layout, and different layouts are triggered by media queries
responding to the screed width.

As with other pages, the images are now delivered using the picture
element to serve the most appropriate sized image, allowing optimization
of bandwidth for smaller devices.