Restore HTML And CSS For Modals
This commit reintroduces the html and css necessary to render the modal
content across the site. Once again this has been refactored into a
mobile first layout and is rendered as an "almost-full-screen" modal for
small phones, or centered column modal for larger phones, with a
breakpoint for a larger modal for tablet+ layouts.