Update Naming Convention
This commit updates the naming of some files and directories in line with CI  advice on the Self Assessment form to use underscores/hyphens in place of spaces. Some construct files have not been edited as the functionality of the website does not depend on them but naming convention will be used for all newly generated files going forward.

[Ticket: X]